each cart will carry a shipping_address_id
the architecture we created allows cart_id to be stored on the shipping
address, or the other way around. this was to facilitate the shipping
of a single order to multiple addresses. that feature has not yet
materialized. another complication is that when a user needs to correct
their cart submission, their address gets saved multiple times.